refactor(canvas): move freeform authoring contract to @posthog/shared
Relocate the freeform canvas authoring contract — the system-prompt builder
(freeformSystemPromptFor), the import whitelist, the starter scaffold, and the
freeform template id — from @posthog/core into @posthog/shared, so @posthog/agent
can consume it (for the canvas_checkout tool) without a core↔agent package cycle.

Pure move: @posthog/core/canvas/{canvasTemplates,freeformSchemas,freeformWhitelist,
freeformStarter} now re-export from @posthog/shared, so existing consumers are
unchanged.

Generated-By: PostHog Code
Task-Id: 2ed89933-091b-446c-af91-5cc72d939b3b
