+++ packages/shared/src/canvas-freeform-whitelist.ts
