+++ packages/shared/tsup.config.ts
@@ -5,6 +5,9 @@ export default defineConfig({
     "src/index.ts",
     "src/agent-platform-types.ts",
     "src/analytics-events.ts",
+    "src/canvas-freeform-prompt.ts",
+    "src/canvas-freeform-starter.ts",
+    "src/canvas-freeform-whitelist.ts",
     "src/constants.ts",
     "src/deeplink.ts",
     "src/dismissalReasons.ts",
